Remove the Parallel namespace
