+++ tests/phpunit/Fixtures/Process/DummyProcessBearer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Fixtures\Process;
 
-use Infection\Process\Runner\Parallel\ProcessBearer;
+use Infection\Process\Runner\ProcessBearer;
 use PHPUnit\Framework\Assert;
 use Symfony\Component\Process\Process;
 use function Safe\sprintf;
