+++ tests/phpunit/Process/Runner/MutationTestingRunnerTest.php
@@ -35,12 +35,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Process\Runner;
 
-use Infection\Process\Runner\Parallel\ProcessRunner;
-use function array_map;
 use ArrayIterator;
-use function count;
-use function get_class;
-use function implode;
 use Infection\AbstractTestFramework\Coverage\TestLocation;
 use Infection\Event\MutationTestingWasFinished;
 use Infection\Event\MutationTestingWasStarted;
@@ -52,7 +47,7 @@ use Infection\PhpParser\MutatedNode;
 use Infection\Process\Builder\MutantProcessBuilder;
 use Infection\Process\MutantProcess;
 use Infection\Process\Runner\MutationTestingRunner;
-use Infection\Process\Runner\Parallel\ParallelProcessRunner;
+use Infection\Process\Runner\ProcessRunner;
 use Infection\Tests\Fixtures\Event\EventDispatcherCollector;
 use Infection\Tests\Mutator\MutatorName;
 use Iterator;
@@ -60,8 +55,12 @@ use PhpParser\Node\Stmt\Nop;
 use PHPUnit\Framework\Constraint\Callback;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use function Safe\sprintf;
 use Symfony\Component\Filesystem\Filesystem;
+use function array_map;
+use function count;
+use function get_class;
+use function implode;
+use function Safe\sprintf;
 
 /**
  * @group integration
